opto/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
gunicorn==20.1.0
Levenshtein==0.24.0
openpyxl==3.1.2
PuLP==2.8.0
python-Levenshtein==0.24.0
pytz==2023.3.post1
rapidfuzz==3.6.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.2.0
